+++ src/nh99/newmodm4.cpp
@@ -23,7 +23,7 @@ double function_minimizer::unrestricted_hess_determinant(void)
   cout << "nvar =" << nvar << endl;
   if (nvar > 0)
   {
-    if ((size_t)nvar != initial_params::nvarcalc())
+    if (nvar != initial_params::nvarcalc())
     {
       cout << "the number of independent variables is wrong in admodel.hes"
          << endl;
